Add a base project, which others can extend from
