+++ source/Halibut.TestUtils.SampleProgram.v5_0_429/Halibut.TestUtils.SampleProgram.v5_0_429.csproj
@@ -13,8 +13,11 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="CliWrap" Version="3.6.1" />
         <PackageReference Include="Halibut" Version="5.0.429" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Halibut.TestUtils.SampleProgram.Base\Halibut.TestUtils.SampleProgram.Base.csproj" />
     </ItemGroup>
 
 </Project>
