+++ source/Halibut.sln
@@ -40,6 +40,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "_build", "..\build\_build.c
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Halibut.TestUtils.SampleProgram.v5_0_429", "Halibut.TestUtils.SampleProgram.v5_0_429\Halibut.TestUtils.SampleProgram.v5_0_429.csproj", "{66014922-3A83-42A2-8C23-D327160175C9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Halibut.TestUtils.SampleProgram.Base", "Halibut.TestUtils.SampleProgram.Base\Halibut.TestUtils.SampleProgram.Base.csproj", "{BBC31D50-5170-49D9-A42D-625768705B7B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -174,6 +176,18 @@ Global
 		{66014922-3A83-42A2-8C23-D327160175C9}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{66014922-3A83-42A2-8C23-D327160175C9}.Release|x86.ActiveCfg = Release|Any CPU
 		{66014922-3A83-42A2-8C23-D327160175C9}.Release|x86.Build.0 = Release|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Debug|x86.Build.0 = Debug|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Release|x86.ActiveCfg = Release|Any CPU
+		{BBC31D50-5170-49D9-A42D-625768705B7B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
